+++ src/api/routes/invitation.router.ts
@@ -7,7 +7,7 @@ import { validateAccess } from '../middlewares/validateAccess.js';
 import { InvitationRepository } from '../persistence/invitation/invitation.repository.js';
 import { ProjectRepository } from '../persistence/project/project.repository.js';
 import { UserProjectRepository } from '../persistence/userProject/userProject.repository.js';
-import { InvitationMailService } from '../services/invitationMail.service.js';
+import { InvitationMailService } from '../services/mail/invitationMail.service.js';
 import { AcceptInvitationUseCase } from '../useCases/invitation/acceptInvitation.useCase.js';
 import { acceptInvitationUseCaseSchema } from '../useCases/invitation/acceptInvitation.useCase.schema.js';
 import { InviteUserUseCase } from '../useCases/invitation/inviteUser.useCase.js';
